testQueryCacheCleanup moved to Nightly.
The test is running over 1 minute. There was an attempt to categorize it as NighlyTest,
but the Category annotation does not work on method level.
